+++ CsvParser/CsvParser/Interfaces/ICsvParser.cs
@@ -8,7 +8,7 @@ namespace CsvParser.Interfaces
     public interface ICsvParser
     {
         CsvConfiguration CsvConfiguration { get; }
-        int Count { get; }
+        long Count { get; }
         bool IsEmpty { get; }
         List<CsvRow> WhereEquals<T>(string fieldName, T fieldValue) where T : IComparable<T>, IConvertible;
         List<CsvRow> WhereGreaterThan<T>(string fieldName, T fieldValue) where T : IComparable<T>, IConvertible;
